Fix mistakes

+++ BikeRental/BikeRental.Domain/Models/Model.cs
@@ -3,47 +3,47 @@
 namespace BikeRental.Domain.Models;
 
 /// <summary>
-/// A class describing the characteristics of the model.
+///     A class describing the characteristics of the model.
 /// </summary>
 public class Model
 {
     /// <summary>
-    /// Unique identifier of the bike model.
+    ///     Unique identifier of the bike model.
     /// </summary>
     public required int Id { get; set; }
 
     /// <summary>
-    /// Wheel size (in inches).
+    ///     Wheel size (in inches).
     /// </summary>
-    public float? WheelSize { get; set; }
+    public double? WheelSize { get; set; }
 
     /// <summary>
-    /// Maximum allowed passenger weight (in kg).
+    ///     Maximum allowed passenger weight (in kg).
     /// </summary>
-    public int? MaxPassengerWeight { get; set; }
+    public double? MaxPassengerWeight { get; set; }
 
     /// <summary>
-    /// Weight of the bike (in kg).
+    ///     Weight of the bike (in kg).
     /// </summary>
-    public float? BikeWeight { get; set; }
+    public double? BikeWeight { get; set; }
 
     /// <summary>
-    /// Type of brakes (e.g., "Disc", "Rim").
+    ///     Type of brakes (e.g., "Disc", "Rim").
     /// </summary>
     public string? BrakeType { get; set; }
 
     /// <summary>
-    /// Model year.
+    ///     Model year.
     /// </summary>
     public int? ModelYear { get; set; }
 
     /// <summary>
-    /// Rental price per hour.
+    ///     Rental price per hour.
     /// </summary>
     public required decimal PricePerHour { get; set; }
 
     /// <summary>
-    /// Type of the bike.
+    ///     Type of the bike.
     /// </summary>
     public required BikeType BikeType { get; set; }
 }
